# Fixed #31275 -- Optimized sql_flush() without resetting sequence on MySQL.