I checked the time taken by `flush` command with 100 tables. See the source code at https://github.com/c-bata/django-fast-mysql-flush

| number of records on each table | before | after `reset_sequences=False` (default) | after `reset_sequences=True` (optional) |
|--- | --- | --- | --- |
| 10|  3.302 sec (+/- 0.076) | 0.517 sec (+/- 0.019) | 3.658 sec (+/- 0.080) |
| 100 | 3.323 sec (+/- 0.047) | 0.575 sec (+/- 0.025) | 3.571 sec (+/- 0.077) |
| 1000 | 3.577 sec (+/- 0.106) | 1.046 sec (+/- 0.029) | 3.813 sec (+/- 0.073) |

Please note that the benchmark script was run on my laptop. So the times should not be taken precisely.


Ticket URL is https://code.djangoproject.com/ticket/31275