@c-bata 2020-03-27
Hmm... I don't know why this change breaks tests like:

![ScreenShot 2020-03-27 22 35 42](https://user-images.githubusercontent.com/5564044/77761383-4e969080-707b-11ea-8880-34755abbac0a.png)

Is it an existing issue?