@charettes 2020-03-27
@c-bata the failure was introduced by your proposed changes but the failure you pointed at seems to be a side effect on another crash

If you look at t[he raw stdout output of the suite execution](https://djangoci.com/job/pull-requests-bionic/database=mysql,label=bionic-pr,python=python3.6/4921/consoleFull) you'll notice that the first failure is 

```
======================================================================
ERROR [0.011s]: test_for_update_after_from (select_for_update.tests.SelectForUpdateTests)
----------------------------------------------------------------------
Traceback (most recent call last):
  File "/home/jenkins/workspace/pull-requests-bionic/database/mysql/label/bionic-pr/python/python3.6/django/db/backends/utils.py", line 84, in _execute
    return self.cursor.execute(sql, params)
  File "/home/jenkins/workspace/pull-requests-bionic/database/mysql/label/bionic-pr/python/python3.6/django/db/backends/mysql/base.py", line 73, in execute
    return self.cursor.execute(query, args)
  File "/home/jenkins/workspace/pull-requests-bionic/database/mysql/label/bionic-pr/python/python3.6/tests/.env/lib/python3.6/site-packages/MySQLdb/cursors.py", line 209, in execute
    res = self._query(query)
  File "/home/jenkins/workspace/pull-requests-bionic/database/mysql/label/bionic-pr/python/python3.6/tests/.env/lib/python3.6/site-packages/MySQLdb/cursors.py", line 315, in _query
    db.query(q)
  File "/home/jenkins/workspace/pull-requests-bionic/database/mysql/label/bionic-pr/python/python3.6/tests/.env/lib/python3.6/site-packages/MySQLdb/connections.py", line 239, in query
    _mysql.connection.query(self, query)
MySQLdb._exceptions.IntegrityError: (1062, "Duplicate entry '1' for key 'person_id'")

The above exception was the direct cause of the following exception:

Traceback (most recent call last):
  File "/home/jenkins/workspace/pull-requests-bionic/database/mysql/label/bionic-pr/python/python3.6/tests/select_for_update/tests.py", line 33, in setUp
    self.person_profile = PersonProfile.objects.create(person=self.person)
  File "/home/jenkins/workspace/pull-requests-bionic/database/mysql/label/bionic-pr/python/python3.6/django/db/models/manager.py", line 82, in manager_method
    return getattr(self.get_queryset(), name)(*args, **kwargs)
  File "/home/jenkins/workspace/pull-requests-bionic/database/mysql/label/bionic-pr/python/python3.6/django/db/models/query.py", line 443, in create
    obj.save(force_insert=True, using=self.db)
  File "/home/jenkins/workspace/pull-requests-bionic/database/mysql/label/bionic-pr/python/python3.6/django/db/models/base.py", line 750, in save
    force_update=force_update, update_fields=update_fields)
  File "/home/jenkins/workspace/pull-requests-bionic/database/mysql/label/bionic-pr/python/python3.6/django/db/models/base.py", line 788, in save_base
    force_update, using, update_fields,
  File "/home/jenkins/workspace/pull-requests-bionic/database/mysql/label/bionic-pr/python/python3.6/django/db/models/base.py", line 891, in _save_table
    results = self._do_insert(cls._base_manager, using, fields, returning_fields, raw)
  File "/home/jenkins/workspace/pull-requests-bionic/database/mysql/label/bionic-pr/python/python3.6/django/db/models/base.py", line 930, in _do_insert
    using=using, raw=raw,
  File "/home/jenkins/workspace/pull-requests-bionic/database/mysql/label/bionic-pr/python/python3.6/django/db/models/manager.py", line 82, in manager_method
    return getattr(self.get_queryset(), name)(*args, **kwargs)
  File "/home/jenkins/workspace/pull-requests-bionic/database/mysql/label/bionic-pr/python/python3.6/django/db/models/query.py", line 1235, in _insert
    return query.get_compiler(using=using).execute_sql(returning_fields)
  File "/home/jenkins/workspace/pull-requests-bionic/database/mysql/label/bionic-pr/python/python3.6/django/db/models/sql/compiler.py", line 1381, in execute_sql
    cursor.execute(sql, params)
  File "/home/jenkins/workspace/pull-requests-bionic/database/mysql/label/bionic-pr/python/python3.6/django/db/backends/utils.py", line 66, in execute
    return self._execute_with_wrappers(sql, params, many=False, executor=self._execute)
  File "/home/jenkins/workspace/pull-requests-bionic/database/mysql/label/bionic-pr/python/python3.6/django/db/backends/utils.py", line 75, in _execute_with_wrappers
    return executor(sql, params, many, context)
  File "/home/jenkins/workspace/pull-requests-bionic/database/mysql/label/bionic-pr/python/python3.6/django/db/backends/utils.py", line 84, in _execute
    return self.cursor.execute(sql, params)
  File "/home/jenkins/workspace/pull-requests-bionic/database/mysql/label/bionic-pr/python/python3.6/django/db/utils.py", line 90, in __exit__
    raise dj_exc_value.with_traceback(traceback) from exc_value
  File "/home/jenkins/workspace/pull-requests-bionic/database/mysql/label/bionic-pr/python/python3.6/django/db/backends/utils.py", line 84, in _execute
    return self.cursor.execute(sql, params)
  File "/home/jenkins/workspace/pull-requests-bionic/database/mysql/label/bionic-pr/python/python3.6/django/db/backends/mysql/base.py", line 73, in execute
    return self.cursor.execute(query, args)
  File "/home/jenkins/workspace/pull-requests-bionic/database/mysql/label/bionic-pr/python/python3.6/tests/.env/lib/python3.6/site-packages/MySQLdb/cursors.py", line 209, in execute
    res = self._query(query)
  File "/home/jenkins/workspace/pull-requests-bionic/database/mysql/label/bionic-pr/python/python3.6/tests/.env/lib/python3.6/site-packages/MySQLdb/cursors.py", line 315, in _query
    db.query(q)
  File "/home/jenkins/workspace/pull-requests-bionic/database/mysql/label/bionic-pr/python/python3.6/tests/.env/lib/python3.6/site-packages/MySQLdb/connections.py", line 239, in query
    _mysql.connection.query(self, query)
django.db.utils.IntegrityError: (1062, "Duplicate entry '1' for key 'person_id'")
```

Since `sql_flush()` is heavily used during tests by `TransactionTestCase` this issue then snowballed to other failures.