@c-bata 2020-03-27
Hi @adamchainz What do you think about changing the default behavior to use `DELETE FROM table_name`?
At least, it is a rare case to create more than 1000 records on unit tests. But `sql_flush()` is used another functions like `flush` management command. So if it's difficult to change the default behavior, it seems to be better to close this pull request.