@c-bata 2020-03-27
The benchmark result of c3c1470 is:

```
(venv) $ python manage.py bench_sql_flush 10
0th elapsed: 2.023
1th elapsed: 1.837
2th elapsed: 1.791
3th elapsed: 1.932
4th elapsed: 1.962
elapsed: 1.909 sec (+/- 0.09399143936388689)
(venv) $ python manage.py bench_sql_flush 100
0th elapsed: 2.149
1th elapsed: 2.059
2th elapsed: 2.042
3th elapsed: 2.169
4th elapsed: 2.163
elapsed: 2.116 sec (+/- 0.06102950355064902)
(venv) $ python manage.py bench_sql_flush 1000
0th elapsed: 2.852
1th elapsed: 2.993
2th elapsed: 3.093
3th elapsed: 2.779
4th elapsed: 2.856
elapsed: 2.915 sec (+/- 0.12635354027361947)
```

This benchmark script was run on my laptop. So the times should not be taken precisely.

https://github.com/c-bata/django-fast-mysql-flush